Add Poetry config to the repository
Poetry has the option to create in-project virtual environments, and VS Code automatically picks up on them if they're created.

This makes getting a new developer up and running on the project a breeze, and 'guarantees' that each dev environment is identical -- albeit at the expense of flexibility for the developer. I'm open to changing this in the future but since it's just me working on this I'm gonna impose my will.