Create a managers subpackage and a base manager class
The current version of the pydev command is designed as a proof of concept,
but isn't really extensible beyond my own personal preferences.

The solution to that, as I see it, is to create a common API for all current
and future supported environment managers to adhere to, which we can then
use as an abstract base class for future environment managers to use.